levers: restore from_values' purity, and two tests that could not do their job
Three fixes, two of them to my own work in the previous commit.

★ `from_values` WAS NO LONGER PURE. It resolved `draft_conf_tau` by calling
`speculative::draft_conf_tau()` inside itself, which reads the real
environment whatever the two closures say. `cargo test` runs this binary's
tests in parallel, so a sibling that set `ATLAS_MTP_DRAFT_CONF=5.0` leaked
into `resolve(&[])` and `the_opt_out_lever_is_on_by_default_and_every_opt_in_is_off`
read `draft_conf_tau: 0.99` out of what should have been a hermetic call.
The tau is now passed IN, alongside `shadow_topk` and `drafter`, which were
already parameters for exactly this reason — and
`externally_resolved_shadow_and_drafter_values_are_carried` now asserts it
carries through. Purity over the closures is the property the whole lever
suite rests on; nothing in this module may read the environment except
`from_env`.

★ A CONTROL THAT PASSED WHEN IT HAD TO FAIL. The source check that the two
halves of the DFlash dump name one flag used `line.contains(FLAG)`. The typo
injected to control it — `ATLAS_DFLASH_DEBUG_DUMP_FULL_TYPO` — contains the
correct name as a PREFIX, so the control went green and the guard was
measuring nothing. It now extracts the quoted variable and compares it
exactly; observed red against the same typo, green on restore. This is the
second guard tonight that a control caught rather than review.

Both new tests are now hermetic. `speculative::parse_draft_conf_tau` is split
from its reader — the same shape as `parse_disable_watchdogs` and
`parse_flag_default_on` — so the clamp is testable without `set_var`. That
matters beyond tidiness: the upper clamp is load-bearing, since an unclamped
`5.0` puts the confidence floor above anything achievable and discards EVERY
draft, turning speculation off with nothing logged.

With the env mutation gone, no test calls `ModelLevers::from_env()` any more,
so its allow list drops back from four entries to three.

`model_levers_tests.rs` was 477 lines, 23 short of the cap, and this wave
would have pushed it over. Split by PROVENANCE rather than by size: the
parent keeps the levers that predate this work, `model_levers_hot_path_tests.rs`
holds the ones lifted out of per-token, per-layer and per-forward
`std::env::var` calls — each pinned with the spelling it had at the site it
came from, because quietly unifying four spellings into one is a
consolidation's characteristic failure. It is a CHILD of `tests`, not a
sibling, so `resolve` is shared rather than copied.

Verified: `cargo test --workspace --locked` 77/77, clippy 0 warnings.
